+++ services/fs/tools.go
@@ -1,0 +1,6 @@
+//go:build tools
+// +build tools
+
+package fs
+
+import _ "go.beyondstorage.io/v5/cmd/definitions"