+++ services/ocios/doc.go
@@ -1,0 +1,6 @@
+/*
+Package ocios provided support for Oracle Object Storage(https://www.oracle.com/cloud/storage/object-storage/).
+*/
+package ocios
+
+//go:generate go run -tags tools go.beyondstorage.io/v5/cmd/definitions service.toml