+++ services/gdrive/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## v0.2.0 - 2021-10-23
+
+### Changed
+
+- feat(services/gdrive): Move services gdrive back (#903)
+- ci(*): Upgrade minimum version to Go 1.16 (#916)
+
+### Upgraded
+
+- build(deps): bump google.golang.org/api in /services/gdrive (#925)
+
 ## v0.1.0 - 2021-09-22
 
 ### Added